AcquisitionManager: split implementation into separate class from MMStudio.
